refactor: refatorando a seção de contatos
